Move reusable template components to components package
