+++ web/src/components/templates/IconPicker/IconPicker.js
@@ -1,9 +1,7 @@
 import React, { useCallback, useState } from "react";
 import PropTypes from "prop-types";
 import { makeStyles } from "@material-ui/styles";
-import SelectableTabs, {
-  SelectableTab,
-} from "../../../components/basic/SelectableTabs";
+import SelectableTabs, { SelectableTab } from "../../basic/SelectableTabs";
 import IconKind from "../../../application/state/templates/IconKind";
 import { useIntl } from "react-intl";
 import StandardIconPicker from "../StandardIconPicker";
