+++ web/src/components/templates/IconPicker/index.js
