+++ web/src/components/templates/IconUploader/IconCropper.js
