+++ web/src/components/templates/IconUploader/IconDropZone.js
