+++ web/src/components/templates/IconUploader/IconUploader.js
