+++ web/src/components/templates/IconUploader/cropImageURL.js
