+++ web/src/components/templates/IconUploader/index.js
