+++ web/src/components/templates/StandardIcon/StandardIcon.js
