+++ web/src/components/templates/StandardIcon/index.js
