+++ web/src/components/templates/StandardIconPicker/IconList.js
