+++ web/src/components/templates/StandardIconPicker/IconOption.js
