+++ web/src/components/templates/StandardIconPicker/StandardIconPicker.js
