+++ web/src/components/templates/StandardIconPicker/index.js
