+++ web/src/components/templates/StandardIconPicker/useIconQuery.js
