+++ web/src/components/templates/TemplateIcon/CustomIcon.js
