+++ web/src/components/templates/TemplateIcon/TemplateIcon.js
