+++ web/src/components/templates/TemplateList/TemplateExamplePreview.js
