+++ web/src/components/templates/TemplateList/TemplateGalleryDialog.js
