+++ web/src/components/templates/TemplateList/TemplateHeader.js
@@ -8,8 +8,8 @@ import ExpandMoreOutlinedIcon from "@material-ui/icons/ExpandMoreOutlined";
 import ChevronRightOutlinedIcon from "@material-ui/icons/ChevronRightOutlined";
 import MoreVertOutlinedIcon from "@material-ui/icons/MoreVertOutlined";
 import IconButton from "@material-ui/core/IconButton";
-import Spacer from "../../../components/basic/Spacer";
-import Button from "../../../components/basic/Button";
+import Spacer from "../../basic/Spacer";
+import Button from "../../basic/Button";
 import TemplateTitle from "./TemplateTitle";
 import TemplateIconPreview from "./TemplateIconPreview";
 import usePopup from "../../../lib/hooks/usePopup";
