+++ web/src/components/templates/TemplateList/TemplateIconPreview.js
